Moved move_container impl to its own module
Had to keep the hook in the main code, but that's not that big of a
deal. Just how the private/public stuff is going to have to work
