+++ src/layout/layout_tree/mod.rs
@@ -1,2 +1,3 @@
 mod layout;
 pub mod tree;
+pub mod movement;
